Issues #176 and #177 Solve API arguments being compatible between contexts
Former-commit-id: 1947d4d78ad2e7333d989e800bc4ff2b45960bfc [formerly ce5e9522fa539b90cd8c4fd4463fcaeb7b2a1127]
Former-commit-id: 56747d72af5bce0901af630154c1edcb63e86477
Former-commit-id: 8b43e2b9fa3d9bda2fb9f41c88ddbb4b09a4d8be
Former-commit-id: 16d097ac5b74a1f93be6ce30288358bf48759b96

+++ Lib/pagebot/elements/views/siteview.py
@@ -65,7 +65,7 @@ class SiteView(HtmlView):
         self.googleAdsAccount = None
         self.googleAnalyticsId = None
     
-    def copyResources(self, path):
+    def copyResources(self, path, removeExisting=False):
         """If self.resourcePaths are defined, then copy them into the destiation path.
         If the resources already exist, then delete them before copy.
         """
@@ -77,7 +77,7 @@ class SiteView(HtmlView):
                     dstPath += resourcePath.split('/')[-1] + '/'
                 if self.verbose:
                     print('[%s.build] Copy %s --> %s' % (self.__class__.__name__, resourcePath, dstPath))
-                if os.path.exists(dstPath):
+                if removeExisting and os.path.exists(dstPath):
                     # Safety check, only run on relative paths
                     #assert dstPath.startswith('/tmp/') or not dstPath.startswith('/'), ('Path must be relative: "%s"' % dstPath) 
                     shutil.rmtree(dstPath)
